Rename spell checking functions
