+++ src/preload.js
@@ -5,7 +5,7 @@ import setupJitsiPreload from './preload/jitsi';
 import setupLinksPreload from './preload/links';
 import setupNotificationsPreload from './preload/notifications';
 import setupSidebarPreload from './preload/sidebar';
-import setupSpellcheckingPreload from './preload/spellchecking';
+import setupSpellcheckingPreload from './preload/spellChecking';
 import setupTitleChangePreload from './preload/titleChange';
 import setupUserPresencePreload from './preload/userPresence';
 import { setupI18next } from './i18n';
